refactor: move skill install helpers out of cli
